🚚 Move: setting폴더 재구성
